+++ spec/spec_helper.rb
@@ -1,6 +1,3 @@
 # frozen_string_literal: true
 
-require 'bundler/setup'
-require 'rspec'
-require 'omniauth'
 require 'omniauth-tanita'