+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # 0.2.0
 
-* set default `provider_ignores_state` to true
+- set default `provider_ignores_state` to true
 
 # 0.1.1
 
-* Update dependent library 'tanita-api-client' to v0.2.3
+- Update dependent library 'tanita-api-client' to v0.2.3
 
 # 0.1.0
 
-* Initial release
+- Initial release